Updated Code Structure
