+++ src/components/Feed/FeedFrame/styles/FeedFrame.module.css
@@ -1,0 +1,41 @@
+@import "../../../../Global.css";
+
+
+.feedContainer {
+  border-radius: 13px;
+  background-color: rgba(12, 29, 44, 1);
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  z-index: 3;
+  display: flex;
+  flex-wrap: wrap;
+  position: relative;
+  top:0;
+  width:43rem;
+  background-color: lightpink;
+  height: 88.5vh;
+  flex-direction: row;
+  padding: 1rem 1rem 1rem 1rem;
+  /* background-color:firebrick;
+  opacity: .5; */
+  overflow: auto;
+
+}
+
+.feedTitle {
+  color: #fff;
+  font: 500 32px Poppins, sans-serif;
+  text-align: left;
+  /* background-color:aqua; */
+  width:100%;
+  /* background-color: lightsalmon; */
+  height: fit-content;
+  text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  margin-bottom: 10px;
+}
+
+@media (max-width: 991px) {
+  .feedTitle {
+    max-width: 100%;
+  }
+}
+
