+++ src/components/Feed/FeedPost/styles/FeedFrame.module.css
@@ -1,30 +1,30 @@
-@import "../../../global.css";
+@import "../../../../Global.css";
 
 
-.feedContainer {
+.postContainer {
   border-radius: 13px;
   background-color: rgba(12, 29, 44, 1);
   box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
   z-index: 3;
   display: flex;
   flex-wrap: wrap;
-  position: relative;
   top:0;
-  width:40.8rem;
-  height: 88.5vh;
+  width:rem;
+  height: 73vh;
   flex-direction: row;
   padding: 1rem 1rem 1rem 1rem;
+  /* background-color: lime; */
   /* background-color:firebrick;
   opacity: .5; */
   overflow-y: auto;
 }
 
-.feedContainer::-webkit-scrollbar {
+.postContainer::-webkit-scrollbar {
   width: 10px;
-  height: 10px;
+  height: 100px;
 }
 
-.feedContainer::-webkit-scrollbar-thumb {
+.postContainer::-webkit-scrollbar-thumb {
   border-radius: 13px;
   border: 1px solid rgba(133, 133, 133, 1);
 }
@@ -51,10 +51,10 @@
   display: flex;
   gap: 15px;
   width: 650px;
-  width: 41rem;
+  width: 3.8rem;
   height: 78vh;
   flex-direction: column;
-  /* background-color: yellow; */
+  background-color: yellow;
   overflow-y: auto;
   padding-right: 10px;
   box-sizing:border-box;
@@ -84,7 +84,7 @@
   width: 100%;
   height: 300.06px;
   gap:0px;
-  /* background-color: purple; */
+  background-color: purple;
   flex-direction: row;
   justify-content:baseline;
   flex-grow: 1;
