+++ src/components/MainFrame/BaseComponent/styles/BaseComponent.module.css
@@ -1,4 +1,4 @@
-@import url("../../../global.css");
+@import url("../../../../Global.css");
 
 .base {
     display: flex;
@@ -22,7 +22,6 @@
     position: relative;
     background-color: var(--color-background);
     box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-    background-color: lightgreen;
     opacity: 0.5;
     display: flex;
     flex-direction: column;
