+++ src/components/MainFrame/EventBar/styles/EventBar.module.css
@@ -1,10 +1,10 @@
-@import "../../../global.css";
+@import "../../../../Global.css";
 
 .eventsContainer {
   justify-content: center;
   /* background-color: red; */
   display: flex;
-  width: 14.86rem;
+  width: 13.86rem;
   height: 100%;
   flex-direction: column;
   font-size: 32px;
