+++ src/components/MainFrame/EventBar/styles/EventCard.module.css
@@ -1,4 +1,4 @@
-@import "../../../global.css";
+@import "../../../../Global.css";
 
 .eventCard {
   border-radius: 8px;
