+++ src/microInteraction/Load/Load.jsx
@@ -1,0 +1,9 @@
+// import load from './styles/load.module.scss'
+
+export default function Loading() {
+  return (
+    <div>
+      <h1>Loading...</h1>
+    </div>
+  )
+}
